Completed Coding train Mandlebrot project.
completed the coding train videos mandlebrot project. All of the necessary files and imports are included. take care to make sure any unused code is deleted as it can lead to the file no running. this is a good launch point for creating our own mandlebrot set calculator and by extension multibrot and julia.