fixed file import. added button for easy switching. removed loop
Different Mandelbrot sets can be switched between with ease. the loop has been removed so it does not update automatically with the sliders. the draw function has been alterred to be a doit function in order to prevent p5 from automatically running the code before the user can input a value for d. the code runs on button press.